pyproject.toml
[project]
name = "chute"
version = "0.1.0"
description = "Internal merge-train bot for a single monorepo"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "pydantic-settings>=2.8.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.2",
    "ty>=0.0.1a14",
]

[project.scripts]
chute = "chute.main:run"
chute-dev = "chute.main:run_dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chute"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
